// src/app/shared/Models/report-setup.ts
export class ReportSetup {

    Id: Number;
    ClassSectionMediumId: Number;
    ExecellentFrom: Number;
    ExcellecntTo: Number;
    GoodFrom: Number;
    GoodTo: Number;
    SatisfactoryFrom: Number;
    SatisfactoryTo: Number;
    PoorFrom: Number;
    PoorTo: Number;
    NotEligibleFrom: Number; 
    NotEligibleTo: Number;
    Selected: boolean;
    Name: string;
       constructor ( ) {
           this.Id = -1;
           
       }
   }